feat(Signal forwarding in threads)
Adds the ability for hull to forward signals. This changes hull to
effectively have 3 threads:

1. A signal handler thread that emits to a channel on signal trap
2. A process runner thread that emits to a channel on process exit
3. The main thread that waits for changes from channels

A list of signals are trapped and forwarded.
hull will only exit on child exit.
